fix: 앨범조회api
